requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
boto3==1.34.53
botocore==1.34.53
certifi==2024.2.2
deprecation==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gotrue==2.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
jmespath==1.0.1
packaging==23.2
postgrest==0.16.1
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
realtime==1.0.2
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
storage3==0.7.0
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
websockets==11.0.3
